Done implementing the load button and its logic
Was similar to save, instead used parse for str to obj conversion, and also removed defaultChecks in checkbos, which were causing the error of not displaying the loaded state, even though when i did console.log it was showing. instead used checked which will be checking the current state of these checkboxes. A bit learning for me i guess :)